few more more typos
Please take a quick look and let me know if any of my changes don't make sense or change your writing in any undesirable way. I don't think there is an issue, but I'd like to be sure due to the number of adjustments I have proposed.

Again, this was a really great article. I learned so much!